Improve test maintainability
