+++ programs/bpf/rust/invoked/src/lib.rs
@@ -1,4 +1,4 @@
 //! @brief Example Rust-based BPF program that issues a cross-program-invocation
 
-pub mod instruction;
+pub mod instructions;
 pub mod processor;
