+++ programs/bpf/rust/realloc/Cargo.toml
@@ -10,7 +10,8 @@ documentation = "https://docs.rs/solana-bpf-rust-realloc"
 edition = "2018"
 
 [features]
-custom-heap = []
+default = ["program"]
+program = []
 
 [dependencies]
 solana-program = { path = "../../../../sdk/program", version = "=1.8.0" }
